Redirect main TO products direct to TO github repo
cc @cooperl09 @jaiswalp -- note that previously the TO PURLs were resolving to the nightly OBO central build at berkeley. This gives you direct control of releases. If you follow the instructions in the README you can make the releases appear automatically. We'll want to do this for your other ontologies too