Update config for some taxon-specific life-stages ontologies.
The repository where the HsapDv, MmusDv, OlatDv, and PdumDv ontologies
(among others) are maintained is undergoing an overhaul that will break
the existing PURLs for those ontologies. This commit updates the
corresponding configuration file to make the PURLs point to the new
location of the latest version of those ontologies.