Make sure CL's IRIs for imported files are resolvable.
Add special cases in the CL configuration for the three types of IRIs
that are used in OWL import declarations within the ontology's edit file
(imports, components, and pattern-derived definitions). The
corresponding files are not release artefacts, so if we redirect those
IRIs using the default generic fall-through (which redirects to the
download folder of the latest release) they will end in 404 errors.
Instead, we redirect those IRIs to point to the master branch.

See https://github.com/obophenotype/cell-ontology/issues/3119.